requirements.txt
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
fastapi==0.115.12
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.1.0
motor==3.7.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
psutil==7.0.0
pydantic==2.11.2
pydantic_core==2.33.1
pymongo==4.11.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==5.2.1
requests==2.32.3
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
websocket==0.2.1
websockets==15.0.1
zope.event==5.0
zope.interface==7.2
